requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.2.1
python-telegram-bot==20.8
whitenoise==6.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
